Refactor for blueprints
